Done validation on Upcoming check-ins create form

+++ web/modules/custom/guest_app_custom/src/EventSubscriber/NodeCreateAlterEventSubscriber.php
@@ -88,56 +88,54 @@ class NodeCreateAlterEventSubscriber implements EventSubscriberInterface {
         $entity->field_hotel_id = [$current_user_hotel_id];
       }
 
-      $hotel_id = get_hotel_id();
+      /*$hotel_id = get_hotel_id();
       $room_type_tid = $entity->get('field_upc_user_room_type')->target_id;
       $check_in_start_date = $entity->get('field_user_checkin_checkout_date')->value;
       $check_in_end_date = $entity->get('field_user_checkin_checkout_date')->end_value;
-      $available_qty_data = hotel_room_available_qty($hotel_id,$room_type);
-      $term_load = Term::load($room_type_tid);
+      $available_qty_data = hotel_room_available_qty($hotel_id,$room_type_tid);
+      //$term_load = Term::load($room_type_tid);
 
       if(!empty($available_qty_data)){
         $available_qty = $available_qty_data['remaining_qty'];
       }
+
       //$rooms = blocked_rooms($hotel_id,$room_type,$check_in_start_date,$check_in_end_date);
 
+      $data = array('nid'=>$entity->id(),'room_type'=>$room_type_tid,'check_in_start_date'=>$check_in_start_date,'check_in_end_date'=>$check_in_end_date,'available_qty'=>$available_qty);
+
       $url = 'public://upcoming_check_ins/'.'Hotel_id_'.$hotel_id.'.json';
+      $tempArray = [];
 
-      $data = ['nid'=>$entity->id(),'room_type'=>$term_load->getName(),'check_in_start_date'=>$check_in_start_date,'check_in_end_date'=>$check_in_end_date,'available_qty'=>$available_qty];
-
-      $file = File::create([
-        'filename' => 'Hotel_id_'.$hotel_id.'.json',
-        'uri' => $url,
-        'status' => 1,
-      ]);
-      //$file->save();
-
-      $dir = dirname($file->getFileUri());
-      if (!file_exists($dir)) {
-        mkdir($dir, 0770, TRUE);
-      }
-      file_put_contents($file->getFileUri(), json_encode($data) );
-      $file->save();
-
-      $file_data = json_decode(file_get_contents($file->getFileUri()) );
-
-       $flag = false;
-        foreach ($parsed_json as $key => $value) {
-            if ($value['nid'] == $entity->id() ) {
-                $flag = true;
-                break;
-            }
+      if (file_exists($url)) {
+        $file_data = json_decode(file_get_contents($url) );
+        $flag = false;
+        foreach ($file_data as $key => $value) {
+          if ($value->nid == $entity->id() ) {
+              $flag = true;
+              break;
+          }
         }
-
-        if($flag){
-          $tempArray = json_decode($inp);
+        if(!$flag){
+          $tempArray = $file_data;
           array_push($tempArray, $data);
           $jsonData = json_encode($tempArray);
-          file_put_contents('results.json', $jsonData);
+          file_put_contents($url, $jsonData);
         }
-
-      //kint($file_data);
-
-      //exit;
+      }
+      else{
+        $file = File::create([
+          'filename' => 'Hotel_id_'.$hotel_id.'.json',
+          'uri' => $url,
+          'status' => 1,
+        ]);
+        //$file->save();
+        $dir = dirname($file->getFileUri());
+        if (!file_exists($dir)) {
+          mkdir($dir, 0770, TRUE);
+        }
+        file_put_contents($file->getFileUri(), json_encode(array($data)) );
+        $file->save();
+      }*/
     }
     /*-----upcoming_check_ins------*/
   }
